requirements.txt
absl-py==0.8.0
aiodns==2.0.0
alabaster==0.7.12
altgraph==0.16.1
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.3
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.8.0
astroid==2.2.5
astropy==3.2.1
atomicwrites==1.3.0
attrs==19.1.0
Automat==0.7.0
autopep8==1.4.4
Babel==2.7.0
backcall==0.1.0
backports.functools-lru-cache==1.5
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.1.7
beautifulsoup4==4.8.0
bitarray==1.0.1
bkcharts==0.2
blaze==0.11.3
bleach==3.1.0
bokeh==1.3.4
boto==2.49.0
boto3==1.9.249
botocore==1.12.249
Bottleneck==1.2.1
bs4==0.0.1
cchardet==2.1.4
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
clyent==1.2.2
colorama==0.4.1
comtypes==1.1.7
conda==4.7.12
conda-build==3.18.9
conda-package-handling==1.6.0
conda-verify==3.4.2
constantly==15.1.0
contextlib2==0.6.0
cryptography==2.7
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.13
cytoolz==0.10.0
dask==2.4.0
datashape==0.5.4
decorator==4.4.0
defusedxml==0.6.0
diff-match-patch==20181111
distributed==2.4.0
Django==2.2.6
django-ckeditor==5.9.0
django-crispy-forms==1.9.1
django-formtools==2.2
django-haystack==2.8.1
django-import-export==2.1.0
django-js-asset==1.2.2
django-mdeditor==0.1.16
django-mptt==0.11.0
django-ranged-response==0.2.0
django-reversion==3.0.7
django-simple-captcha==0.5.12
dlib==19.8.1
docutils==0.15.2
echarts-china-cities-pypkg==0.0.9
echarts-china-provinces-pypkg==0.0.3
echarts-countries-pypkg==0.1.6
entrypoints==0.3
et-xmlfile==1.0.1
face-recognition==1.2.3
face-recognition-models==0.3.0
fastcache==1.1.0
filelock==3.0.12
Flask==1.1.1
flask-core==2.9.0
Flask-Cors==3.0.8
Flask-Login==0.4.1
fsspec==0.5.1
future==0.17.1
gast==0.2.2
gensim==3.8.1
gevent==1.4.0
glob2==0.7
google-pasta==0.1.7
graphviz==0.11.1
greenlet==0.4.15
grpcio==1.24.0
h5py==2.9.0
HeapDict==1.0.1
html5lib==1.0.1
httplib2==0.9.2
hyperlink==19.0.0
idna==2.8
imageio==2.5.0
imagesize==1.1.0
imbalanced-learn==0.5.0
imblearn==0.0
importlib-metadata==0.23
incremental==17.5.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
jieba==0.39
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsmin==2.2.2
json5==0.8.5
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.5.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab==1.1.4
jupyterlab-launcher==0.10.5
jupyterlab-server==1.0.6
jupyterthemes==0.20.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==18.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
lesscpy==0.13.0
libarchive-c==2.8
llvmlite==0.29.0
locket==0.2.0
lxml==4.4.1
macholib==1.11
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
mock==3.0.5
more-itertools==7.2.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
mysqlclient==1.4.2.post1
navigator-updater==0.2.1
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
nltk==3.4.5
nose==1.3.7
notebook==6.0.1
numba==0.45.1
numexpr==2.7.0
numpy==1.16.3
numpydoc==0.9.1
odfpy==1.4.1
odo==0.5.1
olefile==0.46
opencv-python==4.1.0
openpyxl==2.6.3
opt-einsum==3.0.1
packaging==19.2
pandas==0.25.1
pandocfilters==1.4.2
parsel==1.5.2
parso==0.5.1
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.4
patsy==0.5.1
pefile==2019.4.18
pep8==1.7.1
pickleshare==0.7.5
Pillow==6.1.0
pkginfo==1.5.0.1
pluggy==0.13.0
ply==3.11
prettytable==0.7.2
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.2
psutil==5.6.3
psycopg2==2.8.3
psycopg2-binary==2.8.3
py==1.8.0
py4j==0.10.8.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycares==3.0.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.3
PyDispatcher==2.0.5
pyecharts==1.3.1
pyecharts-jupyter-installer==0.0.3
pyflakes==2.1.1
Pygments==2.4.2
PyHamcrest==1.9.0
PyInstaller==3.4
pylint==2.3.1
PyMySQL==0.9.3
pyodbc==4.0.27
pyOpenSSL==19.0.0
pyparsing==2.4.2
PyQt5==5.10.1
PyQt5-sip==4.19.18
PyQtdesignerplugins==1.1
pyquery==1.4.0
pyreadline==2.1
pyrsistent==0.15.4
PySocks==1.7.1
pyspider==0.3.10
pyStrich==0.8
pytest==5.1.2
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.4.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
pytest-runner==5.1
python-dateutil==2.8.0
pytz==2019.2
PyWavelets==1.0.3
pywin32==223
pywin32-ctypes==0.2.0
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
qrcode==6.1
QtAwesome==0.6.0
qtconsole==4.5.5
QtPy==1.9.0
queuelib==1.5.0
redis==3.3.11
requests==2.22.0
rjsmin==1.1.0
rope==0.14.0
ruamel-yaml==0.15.46
s3transfer==0.2.1
scikit-image==0.13.1
scikit-learn==0.21.3
scipy==1.3.1
Scrapy==1.6.0
scrapy-redis==0.6.8
scrapy-splash==0.7.2
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
sip==4.19.8
six==1.12.0
smart-open==1.8.4
snowballstemmer==1.9.1
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.3
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.8
sqlparse==0.3.0
stanfordcorenlp==3.9.1.1
statsmodels==0.10.1
sympy==1.4
tables==3.5.2
tablib==2.0.0
tb-nightly==1.15.0a20190806
tblib==1.4.0
tensorboard==2.0.0
tensorflow-estimator==2.0.0
tensorflow-gpu==2.0.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tf-estimator-nightly==1.14.0.dev2019080601
toolz==0.10.0
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.2
tushare==1.2.39
Twisted==19.7.0
typed-ast==1.4.0
typing==3.6.4
u-msgpack-python==2.5.1
unicodecsv==0.14.1
urllib3==1.24.2
w3lib==1.21.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wordcloud==1.5.0
wrapt==1.11.2
WsgiDAV==3.0.0
xadmin==2.0.1
xgboost==0.90
xlrd==1.2.0
XlsxWriter==1.2.1
xlwings==0.15.10
xlwt==1.3.0
zhon==1.1.5
zict==1.0.0
zipp==0.6.0
zope.interface==4.6.0
